feat: wire cap 63 end-to-end, add Phase 9 Token Awareness Layer
Cap 63 (openclaw_execute) — full Phase 8 governed execution surface:
- Registry entry, topology override, governor timeout (45s) and dispatch
- GovernorMediator routing for 'morning brief' and 'run <template>'
- OpenClawExecuteExecutor: asyncio.run() bridge to agent runner
- 15 tests in tests/openclaw/test_openclaw_execute_executor.py
- inbox_check gap surfaced in dashboard UI (opacity dimming + tooltip)

Phase 9 Token Awareness Layer:
- Budget gate in governor: hard refusal for caps {16,48,49,50,55,56,63} at budget limit
- Budget snapshot attached to result.data for all network-cap results
- WebSocket token_budget_update event emitted after budget-gated actions
- GET /api/token/budget endpoint returns live usage snapshot
- Frontend token-budget-bar: amber warning / red limit banner in chat UI
- 10 tests in tests/phase9/test_token_awareness.py

Pre-existing test fixes (6 categories):
- phase42: combined brain_server + session_handler source for turn-count/deep-mode checks
- STT: patch VOSK_MODEL_PATH to tmp_path so transcribe_bytes check passes
- registry count: 24 → 25 active caps after cap 63 added
- openclaw API: fresh ProviderUsageStore per test to prevent cross-test budget pollution
- governance matrix: regenerated GOVERNANCE_MATRIX_TREE.md with cap 63
- phase45 path: conftest.py autouse fixture injects C:\Nova-Project as candidate in worktree

Worktree path fix:
- tests/phase45/conftest.py: _inject_nova_project_root_candidate fixture ensures
  C:\Nova-Project is always first candidate path in git worktree environments

Additional deliverables:
- src/connectors/email_connector.py: EmailConnector interface stub + registry for inbox_check
- docs/design/NOVA_AUDIT_TODO_2026-03-28.md: cap 63 dual-surface reconciliation + Phase 9 items
- 805 → 815 tests passing, 0 failing
