+++ docs/current_runtime/CURRENT_RUNTIME_STATE.md
@@ -1,6 +1,6 @@
 # NOVA - CURRENT RUNTIME STATE
 
-Runtime Fingerprint: 2364b317b6090ca694e3952ac2f980e0ce0ebf62c41919a0892e7f4adc1e5835
+Runtime Fingerprint: de438cecc9e5cae97b70ac22bb7604a7aab685be853e7e53152f62d6b29afbda
 Generated By: scripts/generate_runtime_docs.py
 
 Authoritative Source: Runtime codebase
@@ -172,9 +172,9 @@ Operational / QA Follow-Through
 - Governor Modules: 6
 - Executors: 19
 - Skills: 5
-- Runtime Surface Hash: 08eeec08729b870c8b9b848afdc0ef6b7325a6f4e2de1a29b6adf122b74a2b80
+- Runtime Surface Hash: 62f99bf3037f330f03a40d71df42fcf80caaddd991a4e59b1e723307b397975b
 
-- Hash: 2364b317b6090ca694e3952ac2f980e0ce0ebf62c41919a0892e7f4adc1e5835
+- Hash: de438cecc9e5cae97b70ac22bb7604a7aab685be853e7e53152f62d6b29afbda
 
 ## Runtime Truth Discrepancies
