+++ docs/current_runtime/RUNTIME_FINGERPRINT.md
@@ -1,6 +1,6 @@
 # RUNTIME_FINGERPRINT
 
-- runtime_surface_hash: 08eeec08729b870c8b9b848afdc0ef6b7325a6f4e2de1a29b6adf122b74a2b80
+- runtime_surface_hash: 62f99bf3037f330f03a40d71df42fcf80caaddd991a4e59b1e723307b397975b
 - enabled_capability_ids_hash: dd065297d2c596e9f6a7648a3cd59bb2e91a3273ee923f5af876f48302cde090
-- runtime_fingerprint_hash: cf3b76d9c3b71ad147ff03b484ca1c72cdbd7684f2a23229ad4cae0ab3db3781
+- runtime_fingerprint_hash: 98f3576cf181163b21432cef9a86228472b2624adceee772b186e2a549a20926
 - phase_marker: Build phase 8
