+++ nova_backend/static/index.html
@@ -85,6 +85,7 @@
 
         <div class="chat-input">
           <div id="thinking-bar" class="thinking-bar" style="display:none" role="status" aria-live="polite"></div>
+          <div id="token-budget-bar" class="token-budget-bar" style="display:none" role="status" aria-live="polite"></div>
           <input id="chat-input" type="text" placeholder="Ask Nova anything..." title="Type a message and press Enter to send." autocomplete="off" spellcheck="false" />
           <button id="deepseek-btn" type="button" aria-label="Ask for a second opinion" title="Ask for a second opinion on the latest message">Second opinion</button>
           <button id="send-btn" type="button" title="Send message (Enter)">Send</button>
