+++ nova_backend/tests/adversarial/test_no_multi_capability_chain.py
@@ -43,12 +43,17 @@ def test_single_request_has_single_action_attempt(monkeypatch):
 
 
 def test_brain_server_does_not_chain_track_cluster_to_track_capability():
-    source = read_text(SRC_ROOT / "brain_server.py")
+    # The track_cluster post-dispatch routing was extracted from brain_server.py into
+    # session_handler.py as part of the session-loop refactor. Check the live location.
+    source = read_text(SRC_ROOT / "websocket" / "session_handler.py")
     assert "if capability_id == 50 and params.get(\"action\") == \"track_cluster\":" in source
     assert "governor.handle_governed_invocation,\n                            52" not in source
 
 
 def test_brain_server_uses_shared_web_planner_not_executor_reference():
-    source = read_text(SRC_ROOT / "brain_server.py")
-    assert "plan_web_open(" in source
-    assert "WebpageLaunchExecutor" not in source
+    # plan_web_open is called from session_handler.py (where cap 17 enrichment happens)
+    # not from brain_server.py directly. The governance contract holds: a shared planner
+    # is used rather than referencing the executor class directly.
+    session_source = read_text(SRC_ROOT / "websocket" / "session_handler.py")
+    assert "plan_web_open(" in session_source
+    assert "WebpageLaunchExecutor" not in session_source
