+++ nova_backend/tests/phase42/test_phase42_turn_count_contract.py
@@ -4,10 +4,16 @@ import re
 
 PROJECT_ROOT = Path(__file__).resolve().parents[3]
 BRAIN_SERVER_PATH = PROJECT_ROOT / "nova_backend" / "src" / "brain_server.py"
+SESSION_HANDLER_PATH = PROJECT_ROOT / "nova_backend" / "src" / "websocket" / "session_handler.py"
+
+
+def _combined_source() -> str:
+    """Return source from brain_server.py + session_handler.py for structural checks."""
+    return BRAIN_SERVER_PATH.read_text(encoding="utf-8") + "\n" + SESSION_HANDLER_PATH.read_text(encoding="utf-8")
 
 
 def test_governed_invocation_branch_increments_turn_count():
-    source = BRAIN_SERVER_PATH.read_text(encoding="utf-8")
+    source = _combined_source()
     match = re.search(
         r"if isinstance\(inv_result, Invocation\):(?P<body>.*?)elif isinstance\(inv_result, Clarification\):",
         source,
@@ -18,7 +24,7 @@ def test_governed_invocation_branch_increments_turn_count():
 
 
 def test_clarification_branch_increments_turn_count():
-    source = BRAIN_SERVER_PATH.read_text(encoding="utf-8")
+    source = _combined_source()
     match = re.search(
         r"elif isinstance\(inv_result, Clarification\):(?P<body>.*?)# --- inv_result is None",
         source,
