+++ nova_backend/tests/phase45/test_policy_center_surface.py
@@ -14,6 +14,7 @@ class _ScriptedWebSocket:
     def __init__(self, messages: list[object]) -> None:
         self._messages = list(messages)
         self.sent_messages: list[dict] = []
+        self.headers = {"host": "testserver", "origin": "http://testserver"}
 
     async def accept(self) -> None:
         return None
